Move database query logic to handler methods
